Create forensic-deploy.yml
This CFN template deploys a VPC, with public, private, reserved & Transit subnets spread across three Availability Zones. It deploys an Internet Gateway, with a default route on the public subnets. It deploys three NAT Gateways (one in each AZ),  and default routes for them in the Transit subnets. It will create a Linux and Windows-based Ec2 instance with default credentials for forensic analysis.